requirements.txt
langchain-openai
langchain-community
langchain-text-splitters
faiss-cpu
gradio
python-dotenv
